[dwmbmpd] Ditch taglib for MPD native tags
I tried using MPD's built-in tag functionality before, but it wouldn't
work correctly. Now I've discovered that mpd_song_get_tag() returns a
pointer which is dependent on mpd_song struct's existence, so the reason
I was getting the error was because I used to free mpd_song from memory
right after obtaining the tag values, which was undefined behavior.