[dwmbcpul] Optimize by minimizing fopen/fclose
The former behavior was to open and close each core freq file, which
resulted in
    CORE_COUNT * UPDATE_FREQUENCY * 2
file openings/closings every second. That's way too much overhead that
can easily be avoided, and from now on, it is.

I haven't been very rigorous testing this, so I'm not entirely sure if
something could break when for example a file was deleted or renamed,
but we're dealing with system device files, so I don't think they can
just disappear easily.