[dwmbmpd] Fix empty queue error and free memory
Got rid of the undefined behavior when no song is present in the queue.
Also, mpd_status will be freed in all cases, not just a few select ones
(this was an error on my part, but with no real impact). Also added a
few stylistic changes - braces around some if statements.