[c-libs/game] Implement objects, instances, event handlers
I've decided to maintain an OVERVIEW.MD file which basically acts as a
comprehensive guide or documentation to the whole library. I do this not
so much to help others (I'm not expecting anyone beside me to use this
library), but to make it easier for myself to remember the goals behind
every implementation and keep track of what's missing.