[c-libs/game] Add RND_gameHandlerAdd, rename some functions
RND_gameHandlerAdd is a function that is used to create a mapping
between object indices and function pointers. This is used to connect
object types to their respective handler functions.

The functions for handlers used to be called "RND_gameHandlers..." (plural).
I changed that to singular, because it's easier to type and looks
better.