[dwmbmpd] Major rewrite, simplification and fixes
1. Removed all dynamic font manipulation with freetype2 and fontconfig
This is because for the past year I've been only using the status bar
with a monospace font, so having advanced machanisms for calculating
string width with a variable-sized width is pointless overhead.

2. Major performance and structural improvements
dwmbmpd was one of the first C programs I've written, and it was riddled
with awkward design choices and badly written code. Many mistakes were
spotted and fixed.

3. Fixed memory safety
valgrind reports no odd behavior

The new strategy involves having a CHAR_WIDTH constant, which denotes
how many pixels wide is each character in the title. This works well, so
long you don't have very weird characters in the title (like kanji,
which sometimes are a tiny bit wider). Still, the visual glitches are
scarce and very subtle.
The new problem is, the CHAR_WIDTH constant may need to be adjusted for
different displays. Soon I'll push a commit that will allow the user to
specify CHAR_WIDTH in an environmental variable, because that seems to
be the easiest way of doing it.